Remove ITrackedReferencesSkipTakeService and add it to TrackedReferencesService
